Revert "refactor: move skip validation check to py_validator (#2013)"
This reverts commit bad24f2543c848202f8fe2c833cd4149fef75d9f.
