+++ crates/blockifier/src/blockifier/stateful_validator.rs
@@ -1,7 +1,9 @@
 use std::sync::Arc;
 
 use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
-use starknet_api::core::{ContractAddress, Nonce};
+use starknet_api::core::Nonce;
+use starknet_api::transaction::TransactionHash;
+use starknet_types_core::felt::Felt;
 use thiserror::Error;
 
 use crate::blockifier::config::TransactionExecutorConfig;
@@ -17,6 +19,7 @@ use crate::state::errors::StateError;
 use crate::state::state_api::StateReader;
 use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::errors::{TransactionExecutionError, TransactionPreValidationError};
+use crate::transaction::objects::TransactionInfo;
 use crate::transaction::transaction_execution::Transaction;
 use crate::transaction::transactions::ValidatableTransaction;
 
@@ -41,19 +44,24 @@ pub type StatefulValidatorResult<T> = Result<T, StatefulValidatorError>;
 /// Manages state related transaction validations for pre-execution flows.
 pub struct StatefulValidator<S: StateReader> {
     tx_executor: TransactionExecutor<S>,
+    max_nonce_for_validation_skip: Nonce,
 }
 
 impl<S: StateReader> StatefulValidator<S> {
-    pub fn create(state: CachedState<S>, block_context: BlockContext) -> Self {
+    pub fn create(
+        state: CachedState<S>,
+        block_context: BlockContext,
+        max_nonce_for_validation_skip: Nonce,
+    ) -> Self {
         let tx_executor =
             TransactionExecutor::new(state, block_context, TransactionExecutorConfig::default());
-        Self { tx_executor }
+        Self { tx_executor, max_nonce_for_validation_skip }
     }
 
     pub fn perform_validations(
         &mut self,
         tx: AccountTransaction,
-        skip_validate: bool,
+        deploy_account_tx_hash: Option<TransactionHash>,
     ) -> StatefulValidatorResult<()> {
         // Deploy account transactions should be fully executed, since the constructor must run
         // before `__validate_deploy__`. The execution already includes all necessary validations,
@@ -63,7 +71,14 @@ impl<S: StateReader> StatefulValidator<S> {
             return Ok(());
         }
 
+        // First, we check if the transaction should be skipped due to the deploy account not being
+        // processed. It is done before the pre-validations checks because, in these checks, we
+        // change the state (more precisely, we increment the nonce).
         let tx_context = self.tx_executor.block_context.to_tx_context(&tx);
+        let skip_validate = self.skip_validate_due_to_unprocessed_deploy_account(
+            &tx_context.tx_info,
+            deploy_account_tx_hash,
+        )?;
         self.perform_pre_validation_stage(&tx, &tx_context)?;
 
         if skip_validate {
@@ -104,6 +119,35 @@ impl<S: StateReader> StatefulValidator<S> {
         Ok(())
     }
 
+    // Check if deploy account was submitted but not processed yet. If so, then skip
+    // `__validate__` method for subsequent transactions for a better user experience.
+    // (they will otherwise fail solely because the deploy account hasn't been processed yet).
+    fn skip_validate_due_to_unprocessed_deploy_account(
+        &mut self,
+        tx_info: &TransactionInfo,
+        deploy_account_tx_hash: Option<TransactionHash>,
+    ) -> StatefulValidatorResult<bool> {
+        let nonce = self
+            .tx_executor
+            .block_state
+            .as_ref()
+            .expect(BLOCK_STATE_ACCESS_ERR)
+            .get_nonce_at(tx_info.sender_address())?;
+        let tx_nonce = tx_info.nonce();
+
+        let deploy_account_not_processed =
+            deploy_account_tx_hash.is_some() && nonce == Nonce(Felt::ZERO);
+        let is_post_deploy_nonce = Nonce(Felt::ONE) <= tx_nonce;
+        let nonce_small_enough_to_qualify_for_validation_skip =
+            tx_nonce <= self.max_nonce_for_validation_skip;
+
+        let skip_validate = deploy_account_not_processed
+            && is_post_deploy_nonce
+            && nonce_small_enough_to_qualify_for_validation_skip;
+
+        Ok(skip_validate)
+    }
+
     fn validate(
         &mut self,
         tx: &AccountTransaction,
@@ -136,17 +180,5 @@ impl<S: StateReader> StatefulValidator<S> {
         )?;
 
         Ok((validate_call_info, tx_receipt))
-    }
-
-    pub fn get_nonce(
-        &mut self,
-        account_address: ContractAddress,
-    ) -> StatefulValidatorResult<Nonce> {
-        Ok(self
-            .tx_executor
-            .block_state
-            .as_ref()
-            .expect(BLOCK_STATE_ACCESS_ERR)
-            .get_nonce_at(account_address)?)
     }
 }
