+++ crates/blockifier/src/blockifier/stateful_validator_test.rs
@@ -1,16 +1,15 @@
-use assert_matches::assert_matches;
 use rstest::rstest;
 use starknet_api::transaction::{Fee, TransactionVersion};
 
 use crate::blockifier::stateful_validator::StatefulValidator;
 use crate::context::BlockContext;
+use crate::nonce;
 use crate::test_utils::contracts::FeatureContract;
 use crate::test_utils::initial_test_state::{fund_account, test_state};
 use crate::test_utils::{CairoVersion, BALANCE};
 use crate::transaction::account_transaction::AccountTransaction;
 use crate::transaction::test_utils::{
-    block_context, create_account_tx_for_validate_test_nonce_0, FaultyAccountTxCreatorArgs,
-    INVALID, VALID,
+    block_context, create_account_tx_for_validate_test_nonce_0, FaultyAccountTxCreatorArgs, VALID,
 };
 use crate::transaction::transaction_types::TransactionType;
 
@@ -64,31 +63,8 @@ fn test_transaction_validator(
     }
 
     // Test the stateful validator.
-    let mut stateful_validator = StatefulValidator::create(state, block_context);
+    let mut stateful_validator = StatefulValidator::create(state, block_context, nonce!(0_u32));
 
-    let result = stateful_validator.perform_validations(tx, false);
-    assert!(result.is_ok(), "Validation failed: {:?}", result.unwrap_err());
-}
-
-#[test]
-fn test_transaction_validator_skip_validate() {
-    let block_context = BlockContext::create_for_testing();
-    let faulty_account = FeatureContract::FaultyAccount(CairoVersion::Cairo1);
-    let state = test_state(&block_context.chain_info, BALANCE, &[(faulty_account, 1)]);
-
-    // Create a transaction that does not pass validations.
-    let tx = create_account_tx_for_validate_test_nonce_0(FaultyAccountTxCreatorArgs {
-        scenario: INVALID,
-        tx_type: TransactionType::InvokeFunction,
-        tx_version: TransactionVersion::THREE,
-        sender_address: faulty_account.get_instance_address(0),
-        class_hash: faulty_account.get_class_hash(),
-        max_fee: Fee(BALANCE),
-        ..Default::default()
-    });
-
-    let mut stateful_validator = StatefulValidator::create(state, block_context);
-    // The transaction validations should be skipped and the function should return Ok.
-    let result = stateful_validator.perform_validations(tx, true);
-    assert_matches!(result, Ok(()));
+    let reuslt = stateful_validator.perform_validations(tx, None);
+    assert!(reuslt.is_ok(), "Validation failed: {:?}", reuslt.unwrap_err());
 }
